Moved "WriteMode" to dal/enums to mirror project directory conventions.
