+++ src/main/java/com/fryrank/dal/enums/WriteMode.java
@@ -1,4 +1,4 @@
-package com.fryrank.dal;
+package com.fryrank.dal.enums;
 
 public enum WriteMode {
     CREATE_IF_ABSENT,
