+++ src/test/java/com/fryrank/dal/PublicUserMetadataDALTests.java
@@ -1,4 +1,5 @@
 package com.fryrank.dal;
+import com.fryrank.dal.enums.WriteMode;
 import com.fryrank.model.PublicUserMetadata;
 import com.fryrank.model.PublicUserMetadataOutput;
 import org.junit.jupiter.api.Test;
